fixed sunrenderer radius not being depacketized correctly; writing out systems in game state log
Former-commit-id: f8e4b9bc8f304383e805a0f425f14f3c0ebe790b

+++ SpaceComponentSystem/Components/Media/SunRenderer.cs
@@ -114,7 +114,7 @@ namespace Space.ComponentSystem.Components
         {
             base.Depacketize(packet);
 
-            Radius = packet.ReadInt32();
+            Radius = packet.ReadSingle();
             SurfaceRotation = packet.ReadVector2();
             PrimaryTurbulenceRotation = packet.ReadVector2();
             SecondaryTurbulenceRotation = packet.ReadVector2();
